miniWatson Stock Q&A System
The idea of developing MiniWatson Stock Q&A System comes from Watson,
which contains the leading technolegy of IBM. By analysing and
understanding the requirements of our users, and searching for all
kinds of information efficiently, we are able to achieve our final goal
– to give the most correct answer to the questions raised by our users.
We wish to solve three problems that most small investors would
encounter. They are, hard to find information, hard to screen
information and hard to analyse information. By doing this, we are able
to help them find what they need in the stock market, and help them
succeed in their trading.